refactor(desktop): split session event handlers and extract tool part logic
